// extensions/vscode/src/contracts/interfaces/plugin-switchboard-strings.interface.ts
export interface IPluginSwitchboardStrings {
	readonly pick: string;
	readonly ours: string;
	readonly yours: string;
	readonly external: string;
	readonly enabled: string;
	readonly disabled: string;
	readonly enable: string;
	readonly disable: string;
	readonly savedRestart: string;
	readonly restart: string;
	readonly noWorkspace: string;
	readonly unavailable: string;
}